Add a Note or inline text
@markjjo I would like to suggest to add a note to clarify that custodian-based holds can NOT be turned OFF and ON via the UI.
It is not explicitly stated in this article that custodian-based hold is not supported to be edited in SCC UI.
Admin should be able to use: Set-CaseHoldPolicy -Enabled $True/$False to make the change.  This came up when a hold failed distribution at first and after it became successful it was still OFF, and had to be turned ON via PowerShell.